+++ java/daikon/inv/sequence/EltIntComparison.java
@@ -57,10 +57,10 @@ class EltIntComparison extends SingleSequence {
   }
 
 
-  public void add_modified(int[] a, int count) {
+  public void add_modified(long[] a, int count) {
     for (int i=1; i<a.length; i++) {
-      int v1 = a[i-1];
-      int v2 = a[i];
+      long v1 = a[i-1];
+      long v2 = a[i];
       if (v1 == v2)
         can_be_eq = true;
       else if (v1 < v2)
