+++ java/daikon/inv/twoSequence/PairwiseLinearBinary.java
@@ -31,7 +31,7 @@ class PairwiseLinearBinary extends TwoSequence {
     return core.format(var1().name, var2().name);
   }
 
-  public void add_modified(int[] x_arr, int[] y_arr, int count) {
+  public void add_modified(long[] x_arr, long[] y_arr, int count) {
     if (x_arr.length != y_arr.length) {
       destroy();
       return;
@@ -40,8 +40,8 @@ class PairwiseLinearBinary extends TwoSequence {
     // int len = Math.min(x_arr.length, y_arr.length);
 
     for (int i=0; i<len; i++) {
-      int x = x_arr[i];
-      int y = y_arr[i];
+      long x = x_arr[i];
+      long y = y_arr[i];
 
       core.add_modified(x, y, count);
       if (no_invariant) {
