+++ java/daikon/inv/twoSequence/TwoSequence.java
@@ -19,7 +19,7 @@ public abstract class TwoSequence extends Invariant {
     return ppt.var_infos[1];
   }
 
-  public void add(int[] v1, int[] v2, int mod_index, int count) {
+  public void add(long[] v1, long[] v2, int mod_index, int count) {
     Assert.assert(! no_invariant);
     Assert.assert((mod_index >= 0) && (mod_index < 4));
     Assert.assert(!finished);
@@ -30,13 +30,13 @@ public abstract class TwoSequence extends Invariant {
     }
   }
 
-  public abstract void add_modified(int[] v1, int[] v2, int count);
+  public abstract void add_modified(long[] v1, long[] v2, int count);
 
   /**
    * By default, do nothing if the value hasn't been seen yet.
    * Subclasses can override this.
    */
-  public void add_unmodified(int[] v1, int[] v2, int count) {
+  public void add_unmodified(long[] v1, long[] v2, int count) {
     return;
   }
